Database File Upload
Database pared down from 300k+ entries to only 1000. This was done for file size & convenience reasons as the original database was almost 7gb in size.
This includes player data, hero choices and mode data, and will be the core of the example project. DB file is around 23mb in size.